+++ lib/client/bundler/proto.js
@@ -3,6 +3,7 @@
 
 var fs = require('fs'),
     path = require('path'),
+    glob = require('glob'),
     shortid = require('shortid'),
     system = require('../system'),
     view = require('../view');
@@ -110,6 +111,7 @@ module.exports = function(ss, bundlers, bundlerById, options) {
       ss.bundler.loadFile(entry,options,formatter,cb, errCb); //TODO call loadFile 'html'
     },
 
+    updateCachedOndemandAssets: UpdateCachedOndemandAssets,
     updateCachedDevAssets: UpdateCachedDevAssets,
 
     clientFilePath: function(rel) {
@@ -361,6 +363,34 @@ module.exports = function(ss, bundlers, bundlerById, options) {
       }
     }
     return ss.bundler.wrapModule(modPath, code);
+  }
+
+  //TODO task for generating the on-demand stuff to target assets dir
+
+  function UpdateCachedOndemandAssets() {
+
+    var files = glob.sync('**/*.js', {cwd: path.join(ss.root,options.dirs.code)}); //TODO need root *.js as well
+    files.forEach(function(rel) {
+      ss.http.cached.route('assets/'+this.client.name+'/code/'+ (rel.replace('\\','/')), null, path.join(ss.root, options.dirs.code, rel) );
+    },this);
+
+    files = glob.sync('**/*.js', {cwd: path.join(ss.root,options.dirs.workers)}); //TODO need root *.js as well
+    files.forEach(function(rel) {
+      ss.http.cached.route('assets/'+this.client.name+'/workers/'+ (rel.replace('\\','/')), null, path.join(ss.root, options.dirs.workers, rel) );
+    },this);
+
+    /*TODO render with bundler
+    var entry = ss.bundler.entryFor('worker',options.dirs.workers.substring(1),path),
+        opts = {
+          //TODO packAssets vs packedAssets
+          compress: options.packAssets
+        };
+
+    ss.bundler.loadFile(entry, opts, null, cb, function(err) {
+      ss.log.clientIssue({name:'unknown','id':'unknown'},options,err,entry);
+      cb('console.log("error","'+ err.userInfoText +'")');
+    });
+    */
   }
 
   function UpdateCachedDevAssets() {
