+++ lib/http/cached.js
@@ -4,6 +4,7 @@
  */
 
 var parseurl = require('parseurl'),
+	debug = require('debug')('http'),
 	send = require('send'),
 	glob = require('glob'),
 	path = require('path');
@@ -83,6 +84,9 @@ module.exports = function cached(ss) {
 			if (filePath) { point.filePath = filePath; }
 		} else {
 			point = new KnownPoint(url, fn, filePath);
+			if (!fn) {
+				debug('added known point %s pointing to static file %s',url,filePath);
+			}
 		}
 	}
 
@@ -95,17 +99,17 @@ module.exports = function cached(ss) {
 		send: sendInternal,
 		set: set,
 
-	    loadStatic: function() {
-        var assets = path.relative(ss.client.dirs.static,ss.client.dirs.assets);
-        var files = glob.sync('**/*',{cwd:ss.client.dirs.static,nodir:true}).filter(function(p) {
-          //TODO exclude directories
-          return p.indexOf(assets) !== 0;
-        });
+    loadStatic: function() {
+      var assets = path.relative(ss.client.dirs.static,ss.client.dirs.assets);
+      var files = glob.sync('**/*',{cwd:ss.client.dirs.static,nodir:true}).filter(function(p) {
+        //TODO exclude directories
+        return p.indexOf(assets) !== 0;
+      });
 
-        files.forEach(function(url) {
-          route(url,null,path.join(ss.client.dirs.static, url));
-        });
-	    },
+      files.forEach(function(url) {
+        route(url,null,path.join(ss.client.dirs.static, url));
+      });
+    },
 
 	    loadAssets: function() {
         var files = glob.sync('**/*',{cwd:ss.client.dirs.assets,nodir:true})
