+++ lib/tasks/defaults.js
@@ -5,7 +5,6 @@ var http = require('http'),
     path = require('path'),
     log = require('../utils/log'),
     view = require('../client/view'),
-    ondemandRoutes = require('../client/serve/ondemand'),
     liveReload = require('./live_reload');
 
 module.exports = function(ss, router, options, orchestrator) {
@@ -51,8 +50,10 @@ module.exports = function(ss, router, options, orchestrator) {
 
     // task: ondemand
     // Listen out for requests to async load new assets
-    ss.defaultTask('serve', function serveOndemand(done) {
-      ondemandRoutes(ss, router, options);
+    ss.defaultTask('serve', ['load-api'], function serveOndemand(done) {
+      ss.bundler.forEach(function(bundler) {
+        bundler.updateCachedOndemandAssets();
+      });
 
       // Send server instance to any registered modules (e.g. console)
       ss.events.emit('server:start', ss.server);
