--- test/unit/client/serve/utils.test.js
@@ -1,37 +1,0 @@
-'use strict';
-
-
-
-describe('client asset serving shared utils', function () {
-
-
-
-    describe('#serve', function () {
-
-
-
-        it('should return the response for a javascript file');
-
-
-
-        it('should return the response for a css file');
-
-
-
-    });
-
-
-
-    describe('#parseUrl', function () {
-
-
-
-        it('should return the pathname, sans querystring');
-
-
-
-    });
-
-
-
-});
